Replaced main controller to use GET.php
