+++ controller/LoginController.php
@@ -26,7 +26,6 @@ class LoginController
 
         //Check if user already logged in - Checking session and cookies
         if ($this->Session->checkIfLoggedIn()) {
-            // echo 'User was logged in..';
             $response->setMessageState(true);
             $response->setMessageString("");
         }
@@ -41,7 +40,6 @@ class LoginController
         // if ($this->checkIfPOST()) {
         if ($this->POST->requestMethodIsPOST()) {
             // CHECK IF ALREADY LOGGED IN -> RESPONSE SHOULD ALREADY BE 'TRUE'
-            // if ($this->LoginView->userWantLogin()) {
             if ($this->POST->userWantToLogin()) {
                 $response = $credentials->validateCredentialFormat();
                 //// SWAPPED THIS LINE BELOW, checks if he's already logged in
@@ -60,7 +58,6 @@ class LoginController
 
                 $this->LayoutView->render($response, $this->LoginView, $this->DateTimeView);
                 // HANDLE LOGOUT && self::$LoginModel->checkIfLoggedIn())
-            // } else if ($this->LoginView->userWantLogout()) {
             } else if ($this->POST->userWantToLogout()) {
                 // Only log out with 'Bye bye!' if the user was logged in
                 if ($this->Session->checkIfLoggedIn()) {
@@ -79,16 +76,6 @@ class LoginController
         }
     }
 
-    // private function userWantLogin()
-    // {
-    //     return isset($_POST['login']);
-    // }
-
-    // private function userWantLogout()
-    // {
-    //     return isset($_POST['logout']);
-    // }
-
     // TESTFUNKTION, WILL BE DELETED WHEN FINISHED
     private function printCredentials(Credentials $credentials)
     {
@@ -99,8 +86,5 @@ class LoginController
         echo $credentials->getCookiePassword();
     }
     // TESTFUNKTION, WILL BE DELETED WHEN FINISHED
-    // private function checkIfPOST()
-    // {
-    //     return $_SERVER['REQUEST_METHOD'] == 'POST';
-    // }
+
 }
