+++ model/Session.php
@@ -28,6 +28,8 @@ class Session
         session_destroy();
     }
 
+    // Add getters for the variable names to be used in LoginView
+
     // public function setName(string $name): void
     // {
     //     $_SESSION[self::$username] = $name;
