+++ view/GET.php
@@ -1,0 +1,13 @@
+<?php
+
+class GET
+{
+    public function userWantToRegister()
+    {
+        if (isset($_GET['register'])) {
+            return true;
+        } else {
+            return false;
+        }
+    }
+}
